requirements.txt
crewai
python-dotenv
google-search-results